+++ src/app/modules/slormancer/constants/common.ts
@@ -17,3 +17,5 @@ export const DELIGHTED_VALUE = 12;
 export const GAME_VERSION = '0.2.152';
 
 export const POISON_DAMAGE_PERCENT = 200;
+
+export const SHIELD_DURATION = 3;