+++ src/app/modules/slormancer/constants/content/data/data-hero-base-stats.ts
@@ -14,6 +14,8 @@ const ALL_HEROES_BASE_STATS = [
     { stat: 'inner_fire_damage_percent', base: -50, perLevel: 0 },
     { stat: 'overdrive_bounce_number_add', base: 2, perLevel: 0 },
     { stat: 'overdrive_damage_percent', base: -40, perLevel: 0 },
+    { stat: 'shield_globe_duration_add', base: 3, perLevel: 0 },
+    { stat: 'shield_globe_value_add', base: 0, perLevel: 3 },
 ];
 
 export const DATA_HERO_BASE_STATS: GameHeroesData<HeroBaseStats> = {