+++ src/app/modules/slormancer/constants/content/data/data-template-mechanic.ts
@@ -2,6 +2,6 @@ import { MechanicType } from '../../../model/content/enum/mechanic-type';
 
 export const DATA_TEMPLATE_MECHANIC: { [key: string]: MechanicType } = {
     '<lucky>': MechanicType.Lucky,
-    '<shield globe>': MechanicType.ShieldGlove,
-    '<shield globes>': MechanicType.ShieldGlove,
+    '<shield globe>': MechanicType.ShieldGlobe,
+    '<shield globes>': MechanicType.ShieldGlobe,
 };