+++ src/app/modules/slormancer/model/content/data/data-mechanic.ts
@@ -1,0 +1,5 @@
+import { AbstractEffectValue } from '../effect-value';
+
+export interface DataMechanic {
+    values: Array<AbstractEffectValue>;
+}