+++ src/app/modules/slormancer/model/content/enum/mechanic-type.ts
@@ -2,5 +2,5 @@ export enum MechanicType {
     InnerFire = 'inner_fire',
     Overdrive = 'overdrive',
     Lucky = 'lucky',
-    ShieldGlove = 'shield_globe'
+    ShieldGlobe = 'shield_globe'
 }