+++ CHANGELOG.md
@@ -47,5 +47,7 @@
 - Cloak of the insatiable now correctly increase bounce number
 
 ## Release 0.0.6 (TBR)
+### Added
+- Added mechanic tooltips
 ### Fixes
 - Fixed an issue with arcane breach configuration and buffs not being applied